Add code to reduce amount of applicable promotions
If we reduce the amount of promotions that can be activated, the
performance when refreshing the cart improves dramatically.

Eventually, this might not be necessary since Solidus#master branch is
updated with https://github.com/solidusio/solidus/pull/4304 and
https://github.com/solidusio/solidus/pull/4296.
Until those land in master and we are able to update to Solidus 3.2, we need to
maintain this fork.